requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0
django-allauth==0.63.6
django-cors-headers==4.7.0
django-crispy-forms==2.2
django-custom-user==1.1
djangorestframework==3.15.0
djangorestframework_simplejwt==5.5.0
filelock==3.13.3
Flask==3.0.3
frozenlist==1.4.1
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
oauthlib==3.2.2
packaging==25.0
phonenumbers==8.13.40
pillow==10.2.0
pip==25.1.1
platformdirs==4.2.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==71.1.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.4
twilio==9.2.3
tzdata==2023.4
urllib3==2.2.2
virtualenv==20.25.1
Werkzeug==3.0.4
yarl==1.9.4
django-extensions==4.1
razorpay==1.4.2
python-slugify==8.0.4 
text-unidecode==1.3
arabic-reshaper==3.0.0
asn1crypto==1.5.1
cryptography==45.0.5 
cssselect2==0.8.0 
html5lib==1.1 
lxml==6.0.0 
oscrypto==1.3.0 
pyHanko==0.29.1 
pyhanko-certvalidator==0.27.0 
pypdf==5.9.0 
python-bidi==0.6.6 
pyyaml==6.0.2 
reportlab==4.4.3 
six==1.17.0 
svglib==1.5.1 
tinycss2==1.4.0 
tzlocal==5.3.1 
uritools==5.0.0 
webencodings==0.5.1 
xhtml2pdf==0.2.17